AI가 오류를 잡아줌. ConsoleHelper는 필요 없나봄..

+++ JkwExtensions/XmlExtensions.cs
@@ -89,7 +89,7 @@ namespace JkwExtensions
             return changedValue;
         }
 
-        public static float GetLong(this XElement xmlNode, string attributeName, long defaultValue = 0)
+        public static long GetLong(this XElement xmlNode, string attributeName, long defaultValue = 0)
         {
             long changedValue = defaultValue;
             if (xmlNode.NodeType != XmlNodeType.Element)
